make conversion more efficient by doing it a step earlier

+++ spectral_cube/spectral_cube.py
@@ -4281,7 +4281,8 @@ class VaryingResolutionSpectralCube(BaseSpectralCube, MultiBeamMixinClass):
 
         # Create the tuple of unit conversions needed.
         factor = cube_utils.bunit_converters(self, unit, equivalencies=equivalencies)
-        factor = np.array(factor)
+        # the unit conversion factor may be float64 by default, so if needed demote (or promote) it first
+        factor = np.array(factor, dtype=self._data.dtype)
 
         # special case: array in equivalencies
         # (I don't think this should have to be special cased, but I don't know
@@ -4291,9 +4292,7 @@ class VaryingResolutionSpectralCube(BaseSpectralCube, MultiBeamMixinClass):
         else:
             data = self._data*factor
 
-        # the unit conversion factor is computed in double precision, which
-        # would otherwise silently inflate, e.g., float32 cubes to float64
-        # (see #995)
+    # possibly redundant: coerce data back into its original dtype
         data = data.astype(self._data.dtype, copy=False)
 
         return self._new_cube_with(data=data, unit=unit)
